+++ library/Makefile
@@ -83,6 +83,7 @@ OBJS_CRYPTO= \
 	     asn1write.o \
 	     base64.o \
 	     bignum.o \
+	     bignum_core.o \
 	     camellia.o \
 	     ccm.o \
 	     chacha20.o \
