+++ library/bignum.c
@@ -39,6 +39,7 @@
 
 #include "mbedtls/bignum.h"
 #include "bignum_internal.h"
+#include "bignum_core.h"
 #include "bn_mul.h"
 #include "mbedtls/platform_util.h"
 #include "mbedtls/error.h"
@@ -111,16 +112,6 @@ void mbedtls_mpi_free( mbedtls_mpi *X )
     X->n = 0;
     X->p = NULL;
 }
-
-
-int mbedtls_mpi_core_calloc( mbedtls_mpi_uint** p, size_t elems )
-{
-    *p = mbedtls_calloc( elems, ciL );
-    if( *p == NULL )
-        return( MBEDTLS_ERR_MPI_ALLOC_FAILED );
-    return( 0 );
-}
-
 
 /*
  * Enlarge to the specified number of limbs
@@ -312,11 +303,6 @@ int mbedtls_mpi_get_bit( const mbedtls_mpi *X, size_t pos )
 
     return( ( X->p[pos / biL] >> ( pos % biL ) ) & 0x01 );
 }
-
-/* Get a specific byte, without range checks. */
-#define GET_BYTE( X, i )                                        \
-    ( ( ( X )[( i ) / ciL] >> ( ( ( i ) % ciL ) * 8 ) ) & 0xff )
-#define GET_BYTE_MPI( X, i ) GET_BYTE( (X)->p, i )
 
 /*
  * Set a bit to a specific value of 0 or 1
@@ -704,97 +690,6 @@ cleanup:
 }
 #endif /* MBEDTLS_FS_IO */
 
-
-/* Convert a big-endian byte array aligned to the size of mbedtls_mpi_uint
- * into the storage form used by mbedtls_mpi. */
-
-static mbedtls_mpi_uint mpi_uint_bigendian_to_host_c( mbedtls_mpi_uint x )
-{
-    uint8_t i;
-    unsigned char *x_ptr;
-    mbedtls_mpi_uint tmp = 0;
-
-    for( i = 0, x_ptr = (unsigned char*) &x; i < ciL; i++, x_ptr++ )
-    {
-        tmp <<= CHAR_BIT;
-        tmp |= (mbedtls_mpi_uint) *x_ptr;
-    }
-
-    return( tmp );
-}
-
-static mbedtls_mpi_uint mpi_uint_bigendian_to_host( mbedtls_mpi_uint x )
-{
-#if defined(__BYTE_ORDER__)
-
-/* Nothing to do on bigendian systems. */
-#if ( __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__ )
-    return( x );
-#endif /* __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__ */
-
-#if ( __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__ )
-
-/* For GCC and Clang, have builtins for byte swapping. */
-#if defined(__GNUC__) && defined(__GNUC_PREREQ)
-#if __GNUC_PREREQ(4,3)
-#define have_bswap
-#endif
-#endif
-
-#if defined(__clang__) && defined(__has_builtin)
-#if __has_builtin(__builtin_bswap32)  &&                 \
-    __has_builtin(__builtin_bswap64)
-#define have_bswap
-#endif
-#endif
-
-#if defined(have_bswap)
-    /* The compiler is hopefully able to statically evaluate this! */
-    switch( sizeof(mbedtls_mpi_uint) )
-    {
-        case 4:
-            return( __builtin_bswap32(x) );
-        case 8:
-            return( __builtin_bswap64(x) );
-    }
-#endif
-#endif /* __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__ */
-#endif /* __BYTE_ORDER__ */
-
-    /* Fall back to C-based reordering if we don't know the byte order
-     * or we couldn't use a compiler-specific builtin. */
-    return( mpi_uint_bigendian_to_host_c( x ) );
-}
-
-static void mpi_bigendian_to_host( mbedtls_mpi_uint * const p, size_t limbs )
-{
-    mbedtls_mpi_uint *cur_limb_left;
-    mbedtls_mpi_uint *cur_limb_right;
-    if( limbs == 0 )
-        return;
-
-    /*
-     * Traverse limbs and
-     * - adapt byte-order in each limb
-     * - swap the limbs themselves.
-     * For that, simultaneously traverse the limbs from left to right
-     * and from right to left, as long as the left index is not bigger
-     * than the right index (it's not a problem if limbs is odd and the
-     * indices coincide in the last iteration).
-     */
-    for( cur_limb_left = p, cur_limb_right = p + ( limbs - 1 );
-         cur_limb_left <= cur_limb_right;
-         cur_limb_left++, cur_limb_right-- )
-    {
-        mbedtls_mpi_uint tmp;
-        /* Note that if cur_limb_left == cur_limb_right,
-         * this code effectively swaps the bytes only once. */
-        tmp             = mpi_uint_bigendian_to_host( *cur_limb_left  );
-        *cur_limb_left  = mpi_uint_bigendian_to_host( *cur_limb_right );
-        *cur_limb_right = tmp;
-    }
-}
-
 /*
  * Import X from unsigned binary data, little endian
  *
@@ -822,18 +717,6 @@ cleanup:
      * input is copied.
      */
     return( ret );
-}
-
-void mbedtls_mpi_core_read_binary( mbedtls_mpi_uint *X, size_t n,
-                                   const unsigned char *buf, size_t buflen )
-{
-    size_t const overhead = ( n * ciL ) - buflen;
-    unsigned char *Xp = (unsigned char*) X;
-    memset( Xp, 0, overhead );
-    if( buflen == 0 )
-        return;
-    memcpy( Xp + overhead, buf, buflen );
-    mpi_bigendian_to_host( X, n );
 }
 
 /*
@@ -901,13 +784,6 @@ int mbedtls_mpi_write_binary_le( const mbedtls_mpi *X,
     }
 
     return( 0 );
-}
-
-void mbedtls_mpi_core_write_binary( const mbedtls_mpi_uint *X,
-                                    unsigned char *buf, size_t buflen )
-{
-    for( size_t i = 0; i < buflen; i++ )
-        buf[buflen - i - 1] = GET_BYTE( X, i );
 }
 
 /*
@@ -1194,90 +1070,6 @@ cleanup:
     return( ret );
 }
 
-/**
- * Helper for mbedtls_mpi subtraction.
- *
- * Calculate l - r where l and r have the same size.
- * This function operates modulo (2^ciL)^n and returns the carry
- * (1 if there was a wraparound, i.e. if `l < r`, and 0 otherwise).
- *
- * d may be aliased to l or r.
- *
- * \param n             Number of limbs of \p d, \p l and \p r.
- * \param[out] d        The result of the subtraction.
- * \param[in] l         The left operand.
- * \param[in] r         The right operand.
- *
- * \return              1 if `l < r`.
- *                      0 if `l >= r`.
- */
-mbedtls_mpi_uint mbedtls_mpi_core_sub( mbedtls_mpi_uint *d,
-                                       const mbedtls_mpi_uint *l,
-                                       const mbedtls_mpi_uint *r,
-                                       size_t n )
-{
-    size_t i;
-    mbedtls_mpi_uint c = 0, t, z;
-
-    for( i = 0; i < n; i++ )
-    {
-        z = ( l[i] <  c );    t = l[i] - c;
-        c = ( t < r[i] ) + z; d[i] = t - r[i];
-    }
-
-    return( c );
-}
-
-mbedtls_mpi_uint mbedtls_mpi_core_sub_int( mbedtls_mpi_uint *d,
-                                           const mbedtls_mpi_uint *l,
-                                           mbedtls_mpi_uint r,
-                                           size_t n )
-{
-    mbedtls_mpi_uint c = 0, t, z;
-
-    z = ( l[0] <  c ); t = l[0] - c;
-    c = ( t < r ) + z; d[0] = t - r;
-
-    for( size_t i = 1; i < n; i++ )
-    {
-        c = ( l[i] < c ); d[i] = l[i] - c;
-    }
-
-    return( c );
-}
-
-mbedtls_mpi_uint mbedtls_mpi_core_add( mbedtls_mpi_uint *d,
-                                       const mbedtls_mpi_uint *l,
-                                       const mbedtls_mpi_uint *r,
-                                       size_t n )
-{
-    mbedtls_mpi_uint c = 0, t;
-    for( size_t i = 0; i < n; i++ )
-    {
-        t  = c;
-        t += l[i]; c  = ( t < l[i] );
-        t += r[i]; c += ( t < r[i] );
-        d[i] = t;
-    }
-    return( c );
-}
-
-mbedtls_mpi_uint mbedtls_mpi_core_lt( const mbedtls_mpi_uint *l,
-                                      const mbedtls_mpi_uint *r,
-                                      size_t n )
-{
-    size_t i;
-    mbedtls_mpi_uint c = 0, t, z;
-
-    for( i = 0; i < n; i++ )
-    {
-        z = ( l[i] <  c ); t = l[i] - c;
-        c = ( t < r[i] ) + z;
-    }
-
-    return( c );
-}
-
 /*
  * Unsigned subtraction: X = |A| - |B|  (HAC 14.9, 14.10)
  */
@@ -1311,19 +1103,11 @@ int mbedtls_mpi_sub_abs( mbedtls_mpi *X, const mbedtls_mpi *A, const mbedtls_mpi
         memset( X->p + A->n, 0, ( X->n - A->n ) * ciL );
 
     carry = mbedtls_mpi_core_sub( X->p, A->p, B->p, n );
+    carry = mbedtls_mpi_core_sub_int( X->p + n, X->p + n, carry, X->n - n );
     if( carry != 0 )
     {
-        /* Propagate the carry to the first nonzero limb of X. */
-        for( ; n < X->n && X->p[n] == 0; n++ )
-            --X->p[n];
-        /* If we ran out of space for the carry, it means that the result
-         * is negative. */
-        if( n == X->n )
-        {
-            ret = MBEDTLS_ERR_MPI_NEGATIVE_VALUE;
-            goto cleanup;
-        }
-        --X->p[n];
+        ret = MBEDTLS_ERR_MPI_NEGATIVE_VALUE;
+        goto cleanup;
     }
 
     /* X should always be positive as a result of unsigned subtractions. */
@@ -1437,38 +1221,6 @@ int mbedtls_mpi_sub_int( mbedtls_mpi *X, const mbedtls_mpi *A, mbedtls_mpi_sint 
     B.p = p;
 
     return( mbedtls_mpi_sub_mpi( X, A, &B ) );
-}
-
-mbedtls_mpi_uint mbedtls_mpi_core_mla( mbedtls_mpi_uint *d, size_t d_len,
-                                       const mbedtls_mpi_uint *s, size_t s_len,
-                                       mbedtls_mpi_uint b )
-{
-    mbedtls_mpi_uint c = 0; /* carry */
-    size_t excess_len = d_len - s_len;
-
-    size_t steps_x8 = s_len / 8;
-    size_t steps_x1 = s_len & 7;
-
-    while( steps_x8-- )
-    {
-        MULADDC_X8_INIT
-        MULADDC_X8_CORE
-        MULADDC_X8_STOP
-    }
-
-    while( steps_x1-- )
-    {
-        MULADDC_X1_INIT
-        MULADDC_X1_CORE
-        MULADDC_X1_STOP
-    }
-
-    while( excess_len-- )
-    {
-        *d += c; c = ( *d < c ); d++;
-    }
-
-    return( c );
 }
 
 /*
@@ -1894,299 +1646,6 @@ int mbedtls_mpi_mod_int( mbedtls_mpi_uint *r, const mbedtls_mpi *A, mbedtls_mpi_
     return( 0 );
 }
 
-/*
- * Fast Montgomery initialization (thanks to Tom St Denis)
- */
-static void mpi_montg_init( mbedtls_mpi_uint *mm, const mbedtls_mpi_uint *N )
-{
-    mbedtls_mpi_uint x, m0 = *N;
-    unsigned int i;
-
-    x  = m0;
-    x += ( ( m0 + 2 ) & 4 ) << 1;
-
-    for( i = biL; i >= 8; i /= 2 )
-        x *= ( 2 - ( m0 * x ) );
-
-    *mm = ~x + 1;
-}
-
-void mbedtls_mpi_core_montmul( mbedtls_mpi_uint *A,
-                               const mbedtls_mpi_uint *B,
-                               size_t B_len,
-                               const mbedtls_mpi_uint *N,
-                               size_t n,
-                               mbedtls_mpi_uint mm,
-                               mbedtls_mpi_uint *T )
-{
-
-    memset( T, 0, (2*n+1)*ciL );
-
-    for( size_t i = 0; i < n; i++, T++ )
-    {
-        mbedtls_mpi_uint u0, u1;
-        /* T = (T + u0*B + u1*N) / 2^biL */
-        u0 = A[i];
-        u1 = ( T[0] + u0 * B[0] ) * mm;
-
-        (void) mbedtls_mpi_core_mla( T, n + 2, B, B_len, u0 );
-        (void) mbedtls_mpi_core_mla( T, n + 2, N, n, u1 );
-    }
-
-    mbedtls_mpi_uint carry, borrow, fixup;
-
-    carry  = T[n];
-    borrow = mbedtls_mpi_core_sub( A, T, N, n );
-    fixup  = carry < borrow;
-    (void) mbedtls_mpi_core_mla( A, n, N, n, fixup );
-}
-
-void mbedtls_mpi_core_add_mod( mbedtls_mpi_uint *X,
-                               mbedtls_mpi_uint const *A,
-                               mbedtls_mpi_uint const *B,
-                               const mbedtls_mpi_uint *N,
-                               size_t n )
-{
-    size_t carry, borrow = 0, fixup;
-    carry  = mbedtls_mpi_core_add( X, A, B, n );
-    borrow = mbedtls_mpi_core_sub( X, X, N, n );
-    fixup  = ( carry < borrow );
-    (void) mbedtls_mpi_core_mla( X, n, N, n, fixup );
-}
-
-int mbedtls_mpi_core_mod( mbedtls_mpi_uint *X,
-                          mbedtls_mpi_uint const *A,
-                          size_t A_len,
-                          const mbedtls_mpi_uint *N,
-                          size_t n,
-                          mbedtls_mpi_uint *RR )
-{
-    int ret = MBEDTLS_ERR_MPI_ALLOC_FAILED;
-    mbedtls_mpi_uint *mempool, *T, *acc, mm, one=1;
-
-    MBEDTLS_MPI_CHK( mbedtls_mpi_core_calloc( &mempool, n+2*n+1) );
-    acc = mempool;
-    T   = mempool + n;
-
-    mpi_montg_init( &mm, N ); /* Compute Montgomery constant */
-    A += A_len; /* Jump to end of A */
-
-    /* The basic idea is the following:
-     * With R = 2^{n*biL}, split A w.r.t. radix R as
-     * A = A0 + R A1 + R^2 A2 + ... = A0 + R(A1 + R(... R(A(n-1) + R*An))...)
-     *
-     * And calculate the iteration X |-> Ai + R*X via combination of
-     * Montgomery multiplication with R^2 and a modular addition. */
-
-    /* Start with top block of A */
-    size_t block_size = A_len % n;
-    if( block_size == 0 )
-        block_size = n;
-
-    A_len -= block_size;
-    A     -= block_size;
-    memset( acc, 0, n*ciL );
-    memcpy( acc, A, block_size * ciL );
-
-    while( A_len >= n )
-    {
-        A_len -= n;
-        A     -= n;
-        /* X |-> R*X mod N via Montgomery multiplication with R^2 */
-        mbedtls_mpi_core_montmul( acc, RR, n, N, n, mm, T );
-        /* Add current block of A */
-        mbedtls_mpi_core_add_mod( acc, acc, A, N, n );
-    }
-
-    /* At this point, we have quasi-reduced the input to the same number
-     * of limbs as the modulus. We get a canonical representative through
-     * two inverse Montomgery multiplications by 1 and R^2.
-     *
-     * TODO: This can be done more efficiently ... one step of Montgomery
-     *       reduction should be enough?
-     *
-     * TODO: Some call-sites seem to be fine with quasi-reduction --
-     *       split this out as a separate function? */
-    mbedtls_mpi_core_montmul( acc, RR, n, N, n, mm, T );
-    mbedtls_mpi_core_montmul( acc, &one, 1, N, n, mm, T );
-
-    memcpy( X, acc, n*ciL ); /* Store result */
-
-cleanup:
-
-    mbedtls_free( mempool );
-    return( ret );
-}
-
-int mbedtls_mpi_core_inv_mod_prime( mbedtls_mpi_uint *X,
-                                    mbedtls_mpi_uint const *A,
-                                    size_t A_len,
-                                    const mbedtls_mpi_uint *P,
-                                    size_t n,
-                                    mbedtls_mpi_uint *RR )
-{
-    int ret = MBEDTLS_ERR_MPI_ALLOC_FAILED;
-    mbedtls_mpi_uint *P2;
-    MBEDTLS_MPI_CHK( mbedtls_mpi_core_calloc( &P2, n ) );
-
-    /* |F_p^x| - 1 = p - 2 */
-    (void) mbedtls_mpi_core_sub_int( P2, P, 2, n );
-    /* Inversion by power: g^|G| = 1 <=> g^{-1} = g^{|G|-1} */
-    MBEDTLS_MPI_CHK( mbedtls_mpi_core_mod( X, A, A_len, P, n, RR ) );
-    MBEDTLS_MPI_CHK( mbedtls_mpi_core_exp_mod( X, X, P, n, P2, n, RR ) );
-
-cleanup:
-
-    mbedtls_free( P2 );
-    return( ret );
-}
-
-/*
- * Sliding-window exponentiation: X = A^E mod N  (HAC 14.85)
- */
-static size_t mpi_exp_mod_get_window_size( size_t Ebits )
-{
-    size_t wsize = ( Ebits > 671 ) ? 6 : ( Ebits > 239 ) ? 5 :
-                   ( Ebits >  79 ) ? 4 : ( Ebits >  23 ) ? 3 : 1;
-
-#if( MBEDTLS_MPI_WINDOW_SIZE < 6 )
-    if( wsize > MBEDTLS_MPI_WINDOW_SIZE )
-        wsize = MBEDTLS_MPI_WINDOW_SIZE;
-#endif
-
-    return( wsize );
-}
-
-int mbedtls_mpi_core_exp_mod( mbedtls_mpi_uint *X,
-                              mbedtls_mpi_uint *A,
-                              const mbedtls_mpi_uint *N,
-                              size_t n,
-                              const mbedtls_mpi_uint *E,
-                              size_t E_len,
-                              const mbedtls_mpi_uint *RR )
-{
-    int ret = MBEDTLS_ERR_ERROR_CORRUPTION_DETECTED;
-    /* heap allocated memory pool */
-    mbedtls_mpi_uint *mempool;
-    /* pointers to temporaries within memory pool */
-    mbedtls_mpi_uint *Wtbl, *Wselect, *temp;
-    /* pointers to table entries */
-    mbedtls_mpi_uint *Wcur, *Wlast, *W1;
-
-    size_t wsize, welem;
-    mbedtls_mpi_uint one = 1, mm;
-
-    mpi_montg_init( &mm, N ); /* Compute Montgomery constant */
-    E += E_len;               /* Skip to end of exponent buffer */
-
-    wsize = mpi_exp_mod_get_window_size( E_len * biL );
-    welem = 1 << wsize;
-
-    /* Allocate memory pool and set pointers to parts of it */
-    const size_t table_limbs   = welem * n;
-    const size_t temp_limbs    = 2 * n + 1;
-    const size_t wselect_limbs = n;
-    const size_t total_limbs   = table_limbs + temp_limbs + wselect_limbs;
-    mempool = mbedtls_calloc( total_limbs, ciL );
-    if( mempool == NULL )
-    {
-        ret = MBEDTLS_ERR_MPI_ALLOC_FAILED;
-        goto cleanup;
-    }
-    Wtbl    = mempool;
-    Wselect = Wtbl    + table_limbs;
-    temp    = Wselect + wselect_limbs;
-
-    /*
-     * Window precomputation
-     */
-
-    /* W[0] = 1 (in Montgomery presentation) */
-    memset( Wtbl, 0, n * ciL ); Wtbl[0] = 1;
-    mbedtls_mpi_core_montmul( Wtbl, RR, n, N, n, mm, temp );
-    Wcur = Wtbl + n;
-    /* W[1] = A * R^2 * R^-1 mod N = A * R mod N */
-    memcpy( Wcur, A, n * ciL );
-    mbedtls_mpi_core_montmul( Wcur, RR, n, N, n, mm, temp );
-    W1 = Wcur;
-    Wcur += n;
-    /* W[i+1] = W[i] * W[1], i >= 2 */
-    Wlast = W1;
-    for( size_t i=2; i < welem; i++, Wlast += n, Wcur += n )
-    {
-        memcpy( Wcur, Wlast, n * ciL );
-        mbedtls_mpi_core_montmul( Wcur, W1, n, N, n, mm, temp );
-    }
-
-    /*
-     * Sliding window exponentiation
-     */
-
-    /* X = 1 (in Montgomery presentation) initially */
-    memcpy( X, Wtbl, n * ciL );
-
-    size_t limb_bits_remaining = 0;
-    mbedtls_mpi_uint window = 0;
-    size_t window_bits = 0, cur_limb;
-    while( 1 )
-    {
-        size_t window_bits_missing = wsize - window_bits;
-
-        const int no_more_bits =
-            ( limb_bits_remaining == 0 ) && ( E_len == 0 );
-        const int window_full =
-            ( window_bits_missing == 0 );
-
-        /* Clear window if it's full or if we don't have further bits. */
-        if( window_full || no_more_bits )
-        {
-            if( window_bits == 0 )
-                break;
-
-            /* Select table entry, square and multiply */
-            mbedtls_ct_table_lookup( (unsigned char*) Wselect,
-                                     (unsigned char*) Wtbl,
-                                     n * ciL, welem, window );
-            for( ; window_bits > 0; window_bits-- )
-                mbedtls_mpi_core_montmul( X, X, n, N, n, mm, temp );
-            mbedtls_mpi_core_montmul( X, Wselect, n, N, n, mm, temp );
-
-            window = 0;
-            continue;
-        }
-
-        /* Load next exponent limb if necessary */
-        if( limb_bits_remaining == 0 )
-        {
-            cur_limb = *--E;
-            E_len--;
-            limb_bits_remaining = biL;
-            continue;
-        }
-
-        const size_t bits_insert =
-            window_bits_missing < limb_bits_remaining ?
-            window_bits_missing : limb_bits_remaining;
-        const size_t bits_mask = ( 1u << bits_insert ) - 1;
-
-        window_bits += bits_insert;
-
-        limb_bits_remaining -= bits_insert;
-        window <<= bits_insert;
-        window |= ( cur_limb >> limb_bits_remaining ) & bits_mask;
-    }
-
-    /* Convert X back to normal presentation */
-    mbedtls_mpi_core_montmul( X, &one, 1, N, n, mm, temp );
-
-    ret = 0;
-
-cleanup:
-
-    mbedtls_free( mempool );
-    return( ret );
-}
-
 int mbedtls_mpi_get_montgomery_constant_unsafe( mbedtls_mpi *RR, mbedtls_mpi const *N )
 {
     int ret = MBEDTLS_ERR_ERROR_CORRUPTION_DETECTED;
@@ -2198,20 +1657,6 @@ int mbedtls_mpi_get_montgomery_constant_unsafe( mbedtls_mpi *RR, mbedtls_mpi con
 
 cleanup:
     return( ret );
-}
-
-void mbedtls_mpi_core_get_montgomery_constant_safe( mbedtls_mpi_uint *RR,
-                                                    mbedtls_mpi_uint const *N,
-                                                    size_t n )
-{
-    /* Start with 2^0=1 */
-    memset( RR, 0, n * ciL );
-    RR[0] = 1;
-
-    /* Repated doubling and modular reduction -- very slow, but compared
-     * to an RSA private key operation it seems acceptable. */
-    for( size_t i=0; i < 2*n*biL; i++ )
-        mbedtls_mpi_core_add_mod( RR, RR, RR, N, n );
 }
 
 int mbedtls_mpi_exp_mod( mbedtls_mpi *X, const mbedtls_mpi *A,
@@ -2268,7 +1713,6 @@ int mbedtls_mpi_exp_mod( mbedtls_mpi *X, const mbedtls_mpi *A,
 
     /* Now input and output have standard size and can be passed to
      * the low-level exponentiation routine. */
-
     mbedtls_mpi_core_exp_mod( X->p, Acopy.p, N->p, N->n, E->p, E->n, RR.p );
 
     /* Correct the sign */
@@ -2422,7 +1866,7 @@ static int mpi_fill_random_internal(
     memset( X->p, 0, overhead );
     memset( (unsigned char *) X->p + limbs * ciL, 0, ( X->n - limbs ) * ciL );
     MBEDTLS_MPI_CHK( f_rng( p_rng, (unsigned char *) X->p + overhead, n_bytes ) );
-    mpi_bigendian_to_host( X->p, limbs );
+    mbedtls_mpi_core_bigendian_to_host( X->p, limbs );
 
 cleanup:
     return( ret );
